backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
beautifulsoup4==4.12.2
duckduckgo-search==3.9.10
groq==0.4.1
python-dotenv==1.0.0
rake-nltk==1.0.6
sumy==0.12.0
diskcache==5.6.3
pydantic==2.13.0
httpx==0.25.2
nltk==3.8.1
slowapi==0.1.9
